backend/pyproject.toml
[project]
name = "heizung"
version = "0.1.0"
description = "Heizungssteuerung Hotel Sonnblick"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic[email]>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "httpx>=0.27",
    "celery>=5.3",
    "redis>=5.0",
    "python-json-logger>=2.0",
    "aiomqtt>=2.3",
    "greenlet>=3.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.2",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RET"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

# Sprint 9.8a: Celery liefert keine Type-Stubs fuer @app.task / @signal.connect-
# Decorators. Mypy meldet sonst untyped-decorator fuer alle Engine-Tasks. Das
# Modul-Override beschraenkt die Suppression auf die Celery-Andockpunkte.
[[tool.mypy.overrides]]
module = ["heizung.celery_app", "heizung.tasks.*"]
disable_error_code = ["untyped-decorator"]
